Bugfix: - fixed a compiling error in image_png.
Feature:
- Enable the NANA_MINGW automatically.

+++ Nana.Cpp11/source/paint/graphics.cpp
@@ -164,7 +164,7 @@ namespace paint
 		{
 			if(impl_ && rhs.impl_ && (this != &rhs))
 				impl_->font_ptr = rhs.impl_->font_ptr;
-			
+
 			return *this;
 		}
 
@@ -255,7 +255,7 @@ namespace paint
 
 				//The object will be delete while dwptr_ is performing a release.
 				drawable_type dw = new nana::detail::drawable_impl_type;
-				//Reuse the old font 
+				//Reuse the old font
 				if(dwptr_)
 				{
 					drawable_type reuse = dwptr_.get();
@@ -538,7 +538,7 @@ namespace paint
 							x += detail::raw_text_extent_size(handle_, str, len).width;
 						}
 
-						str = i; 
+						str = i;
 						while(str != end && (*str == '\t'))
 							++str;
 
@@ -583,7 +583,7 @@ namespace paint
 			if((static_cast<int>(width) > -x) && (static_cast<int>(height) > -y) && width && height && handle_)
 			{
 #if defined(NANA_WINDOWS)
-				::RECT r = {x, y, x + width, y + height};
+				::RECT r = {x, y, static_cast<long>(x + width), static_cast<long>(y + height)};
 				handle_->brush.set(handle_->context, handle_->brush.Solid, color);
 				(solid ? ::FillRect : ::FrameRect)(handle_->context, &r, handle_->brush.handle);
 #elif defined(NANA_X11)
@@ -841,7 +841,7 @@ namespace paint
 				pixbuf.blend(nana::point(s_good_r.x, 0), dst.handle_, d_good_r, fade_rate);
 
 				if(dst.changed_ == false) dst.changed_ = true;
-			}		
+			}
 		}
 
 		void graphics::blend(int x, int y, unsigned width, unsigned height, nana::color_t color, double fade_rate)
